refactor(web): renamed `RoundBuildingCard`
